+++ app/Services/VCard/ExportVCard.php
@@ -269,9 +269,10 @@ class ExportVCard extends BaseService
      */
     private function exportTags(Contact $contact, VCard $vcard)
     {
-        $tags = $contact->getTagsAsString();
-        if (! empty($tags)) {
-            $vcard->CATEGORIES = $tags;
+        if ($contact->tags->count() > 0) {
+            $vcard->CATEGORIES = $contact->tags->map(function ($tag) {
+                return $tag->name;
+            })->toArray();
         }
     }
 }